refactor(editor): table width runtime 분리
